% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{BestFit2PIDLLik_Random}
\alias{BestFit2PIDLLik_Random}
\title{A function rather aimed at developers}
\usage{
BestFit2PIDLLik_Random(
  vX = obs_time,
  vY = obs_prev,
  vW = obs_weig,
  xResBestFit
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
