% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{BestFitPIDLLik_Random}
\alias{BestFitPIDLLik_Random}
\title{A function rather aimed at developers}
\usage{
BestFitPIDLLik_Random(
  vX,
  vY,
  vW,
  maxknots = 3,
  max_FProj = 5,
  ficheck = FALSE,
  frecov
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
