% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{CalcCS}
\alias{CalcCS}
\title{Congenital Syphilis estimation and projection using results from the procedure RunFitSyphilis}
\usage{
CalcCS(
  syphfitobj,
  list_countries = NULL,
  proj_years = 1990:2025,
  min_year = 2011,
  rSyph_preg = 1
)
}
\arguments{
\item{syphfitobj}{an object of class "Syph-fit", output from RunFitSyphilis.}

\item{list_countries}{Vector containing the alpha-numeric ISO codes of countries Congenital Syphilis estimates are to be estimated.}

\item{proj_years}{integer vector of years for which estimates and projections are to be made. By default, this is the sequence 1990:2025}

\item{min_year}{non-zero real number indicating the earliest year for data to be used. This is 2011 by default.}

\item{rSyph_preg}{Ratio of Syphilis prevalence among all women to prevalence among pregnant women. This is 1 by default.}
}
\value{
A list of class "CSProj", of dataframes CongenDataOut, LongCongenDataOutForPlots, RegCSABO, LongRegCSABO, RegDataPrevInc, LongRegDataPrevIncForPlots containing national and regional congenital Syphilis estimates.
}
\description{
Congenital Syphilis estimation and projection using results from the procedure RunFitSyphilis
}
\examples{
Not available
}
