% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{fCountryAnalysis_glob}
\alias{fCountryAnalysis_glob}
\title{A function rather aimed at developers}
\usage{
fCountryAnalysis_glob(
  Nboots = 1000,
  fname.data.file = name.data.file,
  Fmaxknots,
  FB_ProjMax,
  duration.syphilis,
  data.syphilis,
  PopSizeBoth,
  PopSizeMen,
  PopSizeWomen,
  ISO3,
  year_predict,
  zerprev_adj = 1/100,
  MtoFRatio,
  fn_min_year_last_data,
  fn_cores = 4,
  autosavefile = TRUE
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
