% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{FitSegPol2PIDLLik_Random}
\alias{FitSegPol2PIDLLik_Random}
\title{A function rather aimed at developers}
\usage{
FitSegPol2PIDLLik_Random(
  StartXPars,
  VectX,
  ObsY,
  VectWeights,
  incicheck = FALSE,
  recov0,
  uscmax,
  frestrknots,
  scfrac
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
