% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{GetTemplateSyphData}
\alias{GetTemplateSyphData}
\title{Generate Templates for Syphilis data entry forms}
\usage{
GetTemplateSyphData(ctr_iso3 = NULL)
}
\arguments{
\item{ctr_iso3}{Optional, vector of alpha numeric ISO codes of countries for which the Samples should be obtained from. This is NULL by default, implying that the template does not have any data.}
}
\value{
Creates and saves an xlsx file in the working directory
}
\description{
Generate Templates for Syphilis data entry forms
}
\examples{
GetTemplateSyphData()
}
