% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{GMOptimSegPolDF}
\alias{GMOptimSegPolDF}
\title{A function rather aimed at developers}
\usage{
GMOptimSegPolDF(par, fn, tol = 1e-04, maxit = 500)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
