% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{InciSegPol2PID}
\alias{InciSegPol2PID}
\title{A function rather aimed at developers}
\usage{
InciSegPol2PID(nx, nEst_XPars, frestrknots)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
