% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{nLBCheckCalc}
\alias{nLBCheckCalc}
\title{A function rather aimed at developers}
\usage{
nLBCheckCalc(XPars, frestrknots)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
