% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{nUBCheckCalc}
\alias{nUBCheckCalc}
\title{A function rather aimed at developers}
\usage{
nUBCheckCalc(XPars, frestrknots, upb)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
