% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{plot_all_CSProj}
\alias{plot_all_CSProj}
\title{Plot CS estimates by country}
\usage{
plot_all_CSProj(
  xCSProj,
  ctr_iso3,
  fbreaks = c(2010, 2012, 2016, 2020),
  years = 2010:2021
)
}
\arguments{
\item{ctr_iso3}{object of class "CSProj". This is a list containing COngenital Syphilis estimates.}

\item{fbreaks}{A numeric vector of positions for breaks on the y-axis. This set to c(2010,2012,2016,2020) by default.}

\item{years}{A numeric vector of positions for years on the y-axis. This set to 2010:2021 by default.}

\item{ctrname}{Country's alpha-numeric ISO code.}
}
\value{
plot, object from ggplot
}
\description{
Plot CS estimates by country
}
\examples{
Not available
}
