% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{plot_GlobMatSypPrev}
\alias{plot_GlobMatSypPrev}
\title{Plot Global and Regional Maternal Syphilis estimates}
\usage{
plot_GlobMatSypPrev(
  xCSProj,
  proj_years = c(2010:2021),
  fbreaks = c(2010, 2016, 2021)
)
}
\arguments{
\item{xCSProj}{object of class "CSProj". This is a list containing COngenital Syphilis estimates.}

\item{proj_years}{A numeric vercot of ears for which estimates are displayed in charts. This is 2010:2021 by default.}

\item{fbreaks}{A numeric vector of positions for breaks on the y-axis. This set to c(2010,2016,2021) by default.}
}
\value{
plot, object from ggplot
}
\description{
Plot Global and Regional Maternal Syphilis estimates
}
\examples{
Not available
}
