% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{plot_GlobPrevIncR}
\alias{plot_GlobPrevIncR}
\title{Plot Global and Regional Syphilis Incidence and Prevalence Rates estimates}
\usage{
plot_GlobPrevIncR(
  xCSProj,
  proj_years = c(2010:2021),
  fbreaks = c(2010, 2016, 2021)
)
}
\arguments{
\item{xCSProj}{object of class "CSProj". This is a list containing Congenital Syphilis estimates.}

\item{proj_years}{Years for which estimates are displayed in charts.}

\item{fbreaks}{A numeric vector of positions for breaks on the y-axis. This set to c(2010,2016,2021) by default.}
}
\value{
plot, object from ggplot
}
\description{
Plot Global and Regional Syphilis Incidence and Prevalence Rates estimates
}
\examples{
Not available
}
