% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{plot_GlobSyphPerBirth}
\alias{plot_GlobSyphPerBirth}
\title{Plot Global and Regional CS-ABO estimates}
\usage{
plot_GlobSyphPerBirth(xCSProj, proj_years = c(2012, 2016, 2020))
}
\arguments{
\item{xCSProj}{object of class "CSProj". This is a list containing COngenital Syphilis estimates.}

\item{proj_years}{Years for which estimates are displayed in charts.}
}
\value{
plot, object from ggplot
}
\description{
Plot Global and Regional CS-ABO estimates
}
\examples{
Not available
}
