% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{ProjSegPol2PID}
\alias{ProjSegPol2PID}
\title{A function rather aimed at developers}
\usage{
ProjSegPol2PID(XPars0, nVectX, grecov, uscmax, frestrknots, scfrac)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
