% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{RunFitSyphilis}
\alias{RunFitSyphilis}
\title{Syphilis prevalence fits and projection}
\usage{
RunFitSyphilis(
  name.data.file = fin_name.data.file,
  name.popu.file = NULL,
  num_boot = 400,
  in_Fmaxknots = 2,
  Fin_B_ProjMax = 3,
  f_high_risk_adj = 1.1,
  in_year_predict = 1990:2023,
  zero_prev_adj = 1/100,
  Sypdurs = default_syph_durations,
  MtoFRatio = 1,
  FIRST_YEAR = 1990,
  min_year_last_data = 2011,
  list_countries = NULL,
  filter_survey = NULL,
  f_autosavefile = TRUE
)
}
\arguments{
\item{name.data.file}{name of the entry file. This is a data base containing Syphilis' survey and program data. This file needs a special format. Please contact the author (GMahiane@avenirhealth.org) if you have not been provided with such a file yet.}

\item{name.popu.file}{name of the population data entry file. This shall contain populations sizes of the 15-to-49 years old. By default, it is set to NULL, which makes the program use population sizes from Spectrum 2023 (WPP 2022) populations estimates.}

\item{num_boot}{integer, number of boostrap replications.}

\item{in_Fmaxknots}{integer, maximum number of knots allowed for the second order segmented polynomial (spline) used to model the incidence trend. This is set to 3 by default.}

\item{Fin_B_ProjMax}{integer, maximum year allowed for extrapolation for the projection before flat-lining.}

\item{f_high_risk_adj}{non negative real number, adjustment factor for adjustment of prevalence among high risk individuals. The prevalence is deflated if this parameter is lower than 1 and inflated otherwise, so this is expected to be larger than 1. The default value is 1.1}

\item{in_year_predict}{sequence of years incidence and prevalence should be projected. This is set to 1990 to 2023 by default.}

\item{zero_prev_adj}{non-zero real number in the range 0 to 1, to adjust for zero prevalence in the data set. The default value is 1/100}

\item{Sypdurs}{list of three non negative real numbers representing Syphilis durations for three regions of the world. The default value for this is default_syph_durations = list(Sypdur_A=1.28, Sypdur_B=2.42, Sypdur_C=4.13)}

\item{MtoFRatio}{non-zero real number representing an estimate of the male to female ratio. This is one by default.}

\item{FIRST_YEAR}{non-zero real number indicating the earliest year for projections to start. This is 1990 by default.}

\item{min_year_last_data}{non-zero real number indicating the earliest year for data to be used. This is 2011 by default.}

\item{list_countries}{Vector with (alpha) ISO3 codes of countries estimates are to be obtained.}

\item{filter_survey}{Vector with names of data types that are to be included in the analysis.}

\item{f_autosavefile}{boolean indicating whether results should be automatically saved in to a file or not. This is set to FALSE by default}
}
\value{
Creates and saves an xlsx file in the working directory and outputs a list of class "Syph-fit" with the name of that file and the prevalence data used. The created file has a name with the suffix "DDMMYY_zAdj_xx_out", where DDMMYY is the data when the fit was completed, and xx is the prevalence adjustment factor.
}
\description{
Syphilis prevalence fits and projection
}
\examples{
Not available
}
