% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{RunFitSyphilis1}
\alias{RunFitSyphilis1}
\title{A function rather aimed at developers}
\usage{
RunFitSyphilis1(
  name.data.file = fin_name.data.file,
  name.popu.file = NULL,
  num_boot = 400,
  in_Fmaxknots = 2,
  Fin_B_ProjMax = 3,
  f_high_risk_adj = 1.1,
  in_year_predict = 1990:2023,
  zero_prev_adj = 1/100,
  Sypdurs = default_syph_durations,
  MtoFRatio = 1,
  FIRST_YEAR = 1990,
  min_year_last_data = 2011,
  list_countries = NULL,
  filter_survey = NULL,
  rautosavefile = TRUE,
  appendtodefaultDB = FALSE
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
