% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{saveCS}
\alias{saveCS}
\title{Save CS estimates to xlsx file}
\usage{
saveCS(xCSProj, fname = NULL)
}
\arguments{
\item{xCSProj}{object of class "CSProj". This is a list containing Congenital Syphilis estimates.}

\item{fname}{name of the xlsx file to which the results are to be saved.}
}
\value{
boolean, TRUE upon completion of the task
}
\description{
Save CS estimates to xlsx file
}
\examples{
Not available
}
