% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{saveSyphfit}
\alias{saveSyphfit}
\title{Save Syphilis' prevalence and incidence estimates to xlsx file}
\usage{
saveSyphfit(xSyphfit, fname = NULL)
}
\arguments{
\item{xSyphfit}{object of class "Syph-fitj". This is a list Syphilis estimates.}

\item{fname}{name of the xlsx file to which the results are to be saved.}
}
\value{
boolean, TRUE upon completion of the task
}
\description{
Save Syphilis' prevalence and incidence estimates to xlsx file
}
\examples{
Not available
}
