% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegPol2FitPIDLLik_heRandom_cpp.R
\name{SSSegPol2PID_Random}
\alias{SSSegPol2PID_Random}
\title{A function rather aimed at developers}
\usage{
SSSegPol2PID_Random(
  XPars,
  VectX,
  VectY,
  VectWeights,
  grecov,
  uscmax,
  frestrknots,
  scfrac
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
