% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{SyphPlots}
\alias{SyphPlots}
\title{Generate plots for results obtained from the function RunFitSyphilis}
\usage{
SyphPlots(syphfits, countries = NULL, yplot = 2000:2023)
}
\arguments{
\item{syphfits}{list of class "Syph-fit" (containing a file name and the data used to run the fits).}

\item{countries}{vector containing the names of countries for which charts displaying estimates are to be drawn.}
}
\value{
Creates and saves two pdf files in the working directory. The created pdf files have names with the prefix "R_Proj_Prev_", and display Syphilis prevalence data and estimated trends by sex and risk.
}
\description{
Generate plots for results obtained from the function RunFitSyphilis
}
\examples{
Not available
}
