% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{SyphPlotsFF}
\alias{SyphPlotsFF}
\title{Generate plots for results obtained from the function RunFitSyphilis}
\usage{
SyphPlotsFF(
  name.data.file = fin_name.data.file,
  f_high_risk_adj = 1.1,
  in_year_predict = 1990:2023
)
}
\arguments{
\item{name.data.file}{name of the file containing.}

\item{countries}{vector containing the names of countries for which charts displaying estimates are to be drawn.}
}
\value{
Creates and saves two pdf files in the working directory. The created pdf files have names with the prefix "R_Proj_Prev_", and display Syphilis prevalence data and estimated trends by sex and risk.
}
\description{
Generate plots for results obtained from the function RunFitSyphilis
}
\examples{
Not available
}
\keyword{internal}
